+++ rcap/qc_binary_images.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python3
+#!/usr/bin/python3.8
 # -*- coding: utf-8 -*-
 
 
@@ -6,6 +6,7 @@ import argparse
 import logging
 import os
 from datetime import datetime as dt
+from pprint import pformat
 
 import cv2
 import numpy as np
@@ -16,9 +17,11 @@ from __init__ import __version__
 
 
 def parse_options():
-  parser = argparse.ArgumentParser(description='Check tresholded images for pure white slices. Creates CSV of volumes that have more than 90% white pixels.',formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+  parser = argparse.ArgumentParser(description='Check tresholded images for pure white slices. Creates CSV of volumes that have more than a given percentage of white pixels.',formatter_class=argparse.ArgumentDefaultsHelpFormatter)
   parser.add_argument("-v", "--verbose", action="store_true", help="Increase output verbosity")
   parser.add_argument("-V", "--version", action="version", version=f'%(prog)s {__version__}')
+  parser.add_argument("-c", "--cutoff", type=float, default=0.8, help="The minimum percentage of white pixels for a given slice for it to be flagged as invalid.")
+  parser.add_argument("-l", "--list", action="store_true", help="Output TXT files that lists bad binary images produced by segmentation")
   parser.add_argument("path", metavar='PATH', type=str, nargs='+', help='Input directory to process. Must contain folder with thresholded images.')
   args = parser.parse_args()
 
@@ -36,7 +39,7 @@ def parse_options():
   fileHandler = logging.FileHandler(lfp)
   fileHandler.setFormatter(logFormatter)
   fileHandler.setLevel(logging.DEBUG) # always show debug statements in log file
-  # rootLogger.addHandler(fileHandler)
+  rootLogger.addHandler(fileHandler)
 
   consoleHandler = logging.StreamHandler()
   consoleHandler.setFormatter(logFormatter)
@@ -48,8 +51,15 @@ def parse_options():
   return args
 
 
-def find_white_slices(fp, cutoff = 0.9):
+def find_white_slices(fp, cutoff):
   files = sorted([ os.path.join(fp, f) for f in os.listdir(fp) if f.endswith('png') ])
+  logging.debug(pformat(files))
+  logging.debug(os.listdir(fp))
+  logging.debug(f"Thresholded images: {pformat(files)}")
+
+  if len(files) == 0:
+    raise Exception(f"Directory does not contain images: '{fp}'")
+
   flagged_images = []
   flagged_indexes = []
   for f in tqdm(files, total=len(files), desc=f"Processing '{fp}'"):
@@ -61,15 +71,15 @@ def find_white_slices(fp, cutoff = 0.9):
       flagged_indexes.append(i)
   flagged_images = sorted(flagged_images)
   if len(flagged_indexes) > 0:
-    return min(flagged_indexes), max(flagged_indexes)
-  return None, None
+    return min(flagged_indexes), max(flagged_indexes), flagged_images
+  return None, None, None
 
 if __name__ == "__main__":
   args = parse_options()
   # Collect all volumes and validate their metadata
   try:
     # Gather all files
-    args.path = [ f'{p}_thresholded_images' for p in args.path if os.path.isdir(p) ]
+    args.path = [ os.path.realpath(p) for p in args.path if os.path.isdir(p) ]
     args.path = list(set(args.path)) # remove duplicates
     for fp in args.path:
       if not os.path.isdir(fp):
@@ -83,27 +93,34 @@ if __name__ == "__main__":
   else:
     if len(args.path) > 0:
       # For each provided directory...
-      flagged_files = []
-      volume_paths = []
+      failed_volumes = []
+      passed_volumes = []
+      # Get the sub-directories that contain binary images
+      binary_image_folders = []
       for parent_path in args.path:
-        volume_paths.extend( [ os.path.join(parent_path, vp) for vp in os.listdir(parent_path) ] )
-      volume_paths = [ vp for vp in volume_paths if os.path.isdir(vp) ]
-      for fp in tqdm(volume_paths, desc=f"Overall progress"):
+        binary_image_folders.extend( [ os.path.join(parent_path, vp) for vp in os.listdir(parent_path) ] )
+      binary_image_folders = [ vp for vp in binary_image_folders if os.path.isdir(vp) ]
+      for fp in tqdm(binary_image_folders, desc=f"Overall progress"):
         logging.debug(f"Processing '{fp}'")
         # Extract slices for all volumes in provided folder
-        start, end = find_white_slices(fp)
-        if start is not None and end is not None:
-          flagged_files.append((fp, start, end))
+        start, end, flagged_slices = find_white_slices(fp, args.cutoff)
+        if start is not None and end is not None and flagged_slices is not None:
+          failed_volumes.append((fp, start, end))
+          flagged_slice_op = f"{os.path.basename(fp)}.flagged_slices.txt"
+          with open(flagged_slice_op, 'w') as ofp:
+            ofp.write('\n'.join(flagged_slices))
+        else:
+          passed_volumes.append(fp)
 
-      if len(flagged_files) > 0:
-        logging.debug(flagged_files)
-        df = pd.DataFrame.from_records(flagged_files, columns = [ 'path', 'start', 'end' ])
+      if set(passed_volumes) == set(binary_image_folders):
+        logging.info(f"All volumes pass!")
+      else:
+        logging.debug(failed_volumes)
+        df = pd.DataFrame.from_records(failed_volumes, columns = [ 'path', 'start', 'end' ])
         df['volume_name'] = df['path'].apply(os.path.basename)
         ofp = f"{dt.today().strftime('%Y-%m-%d_%H-%M-%S')}_{os.path.splitext(os.path.basename(__file__))[0]}.flagged_volumes.csv"
         df.to_csv(ofp, index=False)
-        logging.info(f"Created '{ofp}'")
-      else:
-        logging.info(f"All volumes look good!")
-
+        logging.debug(pformat(df))
+        logging.info(f"Detected possible incorrect segmentation. Check volumes listed in '{ofp}' for details.")
     else:
       logging.info(f"No volumes supplied.")