+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.5.0 - 2020-04-23
+
+### Added
+
+- Script to downsample point cloud data for quality control version of point cloud data
+- Added cutoff value to input options for binary image segmentation script
+- Detect bit depth of .RAW volumes during `raw2img` processing
+
+### Changed
+
+- Renamed quality control script for binary images
+- Restructure files into bin, src, and package
+
 ## v1.4.1 - 2020-04-14
 
 ### Fixed